+++ src/redturtle/prenotazioni/interfaces.py
@@ -2,6 +2,7 @@
 """Module where all interfaces, events and exceptions live."""
 
 from zope.interface import Interface
+from zope.interface.interfaces import IObjectEvent
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
 
@@ -17,9 +18,33 @@ class ISerializeToPrenotazioneSearchableItem(Interface):
     """Prenotazione searchable item serializer interface"""
 
 
+class ISerializeToRetroCompatibleJson(Interface):
+    """Interface used to cereate the TEMPORARY retrocomattible serializers"""
+
+
+class IBookingReminderEvent(IObjectEvent):
+    """Booking reminder time arrived event"""
+
+
 class IPrenotazioneEmailMessage(Interface):
     """Prenotazione email message"""
 
 
-class ISerializeToRetroCompatibleJson(Interface):
-    """Interface used to cereate the TEMPORARY retrocomattible serializers"""
+class IBookingNotificationSender(Interface):
+    """Booking notification sender"""
+
+
+class IPrenotazioneSMSMessage(Interface):
+    """Prenotazione SMS message adapter"""
+
+
+class IPrenotazioneAPPIoMessage(Interface):
+    """Prenotazione AppIO message adapter"""
+
+
+class IBookingAfterTransitionEvent(IObjectEvent):
+    """Booking After transition email"""
+
+
+class IBookingNotificatorSupervisorUtility(Interface):
+    """Bookign notificator supervisor"""