+++ CHANGES.rst
@@ -13,6 +13,11 @@ Changelog
 
 - Fix upgrade step error #139 
   [mamico]
+- Split booking notification gateways to 'App IO', 'Email' and 'SMS'
+  [folix-01]
+
+- Add booking reminder emails.
+  [folix-01]
 
 
 2.2.5 (2023-12-04)
@@ -47,6 +52,7 @@ Changelog
 
 - Set PrenotazioniFolder and PrenotazioneType as not searchable (types_not_searched).
   [cekk]
+
 - Send ical also for manager notifications.
   [cekk]
 
@@ -84,7 +90,7 @@ Changelog
 
 - Add booking_refuse_message to Prenotazione stringinterp variables.
   [folix-01]
- 
+
 - Extended PrenotaizoniFolder email templates var list.
   [folix-01]